requirements.txt
nonebot2==2.5.0
nonebot-adapter-telegram==0.1.0b20
nonebot-plugin-parser==2.6.3
nonebot-plugin-alconna==0.62.0
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-uninfo==0.11.0
nonebot-plugin-waiter==0.8.1
nonebot-plugin-localstore==0.7.4
nonebot-plugin-status==0.9.0
curl_cffi==0.13.0
httpx==0.28.1
yt-dlp
pillow
fastapi
uvicorn[standard]
arclet-alconna==1.8.44
arclet-alconna-tools==0.7.12
